Add SCSS and compilation stages
This commit adds a bundle of structural SCSS and plumbs it into the rest
of the template.

There was some pain copying across the scss/ dir from inside a JAR as
you cannot treat it as a real directory. I've pulled in some 3rd party
code it iterate JAR files based on base path and then written some glue
to copy them across.

`base-path` was stolen from the source of `->files` inside of leiningen
so should be a reliable way of devining the new projects path.

We rely on sassc rather than running sass on JRuby or something because
that sounds like a right royal pain and a bit of shelling out is far
simpler. There is some overhead of keeping versions of sassc inline
across CI and dev.