Add templating engine
Settled on Selmer as it hits a nice spot between Mustache and going
fully Hiccup. At the moment any file reference has to include the
leading templates/ as I've not configured the `resource-path` in Selmer
to remove this, see: https://github.com/yogthos/Selmer#resource-path.
I'm not sure where would be best to put this line and its fairly uneeded
at the moment.