Option to initialise a git repo in project dir
The reason we do a nil? check in the predicate for ask-until is that
when the shell that `lein new` is running in is non-interactive
read-line will return nil. We want to accept this value as we want no
flag set through CLI options and no value inputted into check to mean
false. If we didn't have this check it would just hang when a
non-interactive running lein new didn't have --git on its command.